refactor(miner/cli): cleanup mess
